# bachelor/third year/Web_and_Software_Architecture/src/doc/api.yaml
openapi: 3.0.0
info:
  title: WasaText API
  description: API for WASAText messaging platform
  version: 1.0.0
...
security: # Apply the security globally to all operations
  - bearerAuth: [] # use the same name as above
paths:
...
  /users/{userId}/groups/{groupId}/members/:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/groupId"
    put:
      tags:
        - group
      summary: Add a user to a group
      description: Allows the user to add another user to an existing group.
      operationId: addToGroup
      requestBody:
        description: User to add to the group
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        '204':
          description: User added successufully
        '400':
          description: Invalid group id or user id
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Group or user not found
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    get:
      tags:
        - group
      summary: Get group members
      description: Returns a list of usernames of users present in the specified group.
      operationId: getGroupMembers
      responses:
        '200':
          description: List of group members
          content:
            application/json:
              schema:
                description: "Schema for get group members"
                type: array
                minItems: 0
                maxItems: 10
                items:
                  $ref: "#/components/schemas/Username"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Group not found or user does not have access to the group
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes: # Define the security scheme type (HTTP bearer)
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
  schemas:
...
    User:
      title: User
      description: "Represents a single User."
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/UserId'
        username:
          $ref: '#/components/schemas/Username'
...
    Username:
      description: Username of the user.
      type: string
      example: "user123"
      minLength: 3
      maxLength: 16
      pattern: "^.*?$"
    UserId:
      description: Unique user identifier
      type: integer
      example: 1
      readOnly: true
...
  parameters:
    userId:
      schema:
        $ref: '#/components/schemas/UserId'
      name: userId
      in: path
      required: true
      description: User id
...
    groupId:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: "Unique identifier for the group."
      name: groupId
      in: path
      required: true
      description: Group id
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
    InternalServerError:
      description: Internal server error